ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
PyPDF2==3.0.1
python-docx==1.1.0
spacy==3.7.2
scikit-learn==1.3.2
requests==2.31.0
